% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples_dataset.R
\docType{data}
\name{JAKi_data}
\alias{JAKi_data}
\title{sample data for analysis}
\format{
An object of class \code{data.frame} with 135 rows and 494 columns.
}
\usage{
JAKi_data
}
\description{
sample data for analysis
}
\keyword{datasets}
