% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metab_list_data.R
\docType{data}
\name{metab_list}
\alias{metab_list}
\title{metabolite list with HMDB ID}
\format{
An object of class \code{data.frame} with 61 rows and 3 columns.
}
\usage{
metab_list
}
\description{
metabolite list with HMDB ID
}
\keyword{datasets}
