% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRMetab_binning.R
\name{NMRMetab_binning}
\alias{NMRMetab_binning}
\title{Binning of NMR spectra}
\usage{
NMRMetab_binning(data, pattern_file)
}
\arguments{
\item{data}{a data.frame. Column as samples and rows as ppm. fits columm 'ppm'. this come from bruker topspin}

\item{pattern_file}{a data.frame with column as 'min_ppm', max_ppm' and 'bin'. this is the pattern file the you created}
}
\description{
Binning of NMR spectra
}
\author{
Michele Fresneda Alarcon
}
