% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRMetab_foldchange.R, R/NMRMetab_univariate.R
\name{NMRMetab_foldchange}
\alias{NMRMetab_foldchange}
\alias{NMRMetab_UnivarTest}
\title{foldchange function}
\usage{
NMRMetab_foldchange(data, groupID, index_col = 3, dividendID, divisorID)

NMRMetab_UnivarTest(
  data,
  groupID,
  index_col = 3,
  paired = F,
  normality = T,
  equal.variance = F
)
}
\arguments{
\item{data}{a data.frame. Column as variable and rows as sample}

\item{groupID}{string. name off the variable containing your grouping details must only have 2 facors eg Healthy vs Diseases}

\item{index_col}{index colum oif first mobservation}

\item{dividendID}{string. top of the division}

\item{divisorID}{string. what group you're dividing by}

\item{paired}{boolean}

\item{normality}{boolean}

\item{equal.variance}{boolean}
}
\description{
This function will take in either a object of class data.frame and a using a second dataframe with a list of all the metabolite identifed, will return the CRS score for all the bins. this could work with other metabolite identifiers; for example bins with the name of the     metabolite, as  log as they are unique. mistakes could happen when greple take NAD and NADH

This function will take in either a object of class data.frame and a using a second dataframe with a list of all the metabolite identifed, will return the CRS score for all the bins. this could work with other metabolite identifiers; for example bins with the name of the     metabolite, as  log as they are unique. mistakes could happen when greple take NAD and NADH
}
\author{
Michele Fresneda Alarcon

Eva Caamano Gutierrez

Arturas Grauslys
}
