% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRMetab_PCA_plot.R
\name{NMRMetab_PCA_loading_plot}
\alias{NMRMetab_PCA_loading_plot}
\title{PCA analysis loading plot}
\usage{
NMRMetab_PCA_loading_plot(data, index_col = 2, pcs = c(1, 2))
}
\arguments{
\item{data}{a data.frame. Column as variable and rows as sample}

\item{index_col}{index colum oif first mobservation}

\item{pcs}{string. name off the fvariable containing your grouping details}
}
\description{
This function will take in either a object of class data.frame and a using a second dataframe with a list of all the metabolite identifed, will return the CRS score for all the bins. this could work with other metabolite identifiers; for example bins with the name of the     metabolite, as  log as they are unique. mistakes could happen when greple take NAD and NADH
}
\author{
Michele Fresneda Alarcon
}
