% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRMetab_PCA_plot.R
\name{NMRMetab_PCA_plot}
\alias{NMRMetab_PCA_plot}
\title{PCA analysis. plot}
\usage{
NMRMetab_PCA_plot(
  data,
  groupID,
  index_col = 2,
  elipses = F,
  pcs = c(1, 2),
  size_point = 3
)
}
\arguments{
\item{data}{a data.frame. Column as variable and rows as sample}

\item{groupID}{string. name off the fvariable containing your grouping details}

\item{index_col}{index colum oif first mobservation}

\item{elipses}{number of components}

\item{pcs}{vector with integers only. principal components to plot default 1, 2}

\item{size_point}{double. control the size of the points in the plot}
}
\description{
This function will take in either a object of class data.frame and a using a second dataframe with a list of all the metabolite identifed, will return the CRS score for all the bins. this could work with other metabolite identifiers; for example bins with the name of the     metabolite, as  log as they are unique. mistakes could happen when greple take NAD and NADH
}
\author{
Michele Fresneda Alarcon
}
