% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRMetab_plots.R
\name{NMRMetab_plot_binned}
\alias{NMRMetab_plot_binned}
\title{plot NMR_data}
\usage{
NMRMetab_plot_binned(binned_data, index_col = 2, group_var = "sampleID")
}
\arguments{
\item{binned_data}{a data.frame. Column as variable and rows as sample}

\item{index_col}{integer. index number of the colum with the firsst metabllite measurement}

\item{group_var}{string. the name of the grouping vairable. defails is sampleID. change appropiately}
}
\description{
plot binned spectra
}
\author{
Michele Fresneda Alarcon
}
