% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRMetab_plots.R
\name{NMRMetab_plot_raw_with_bins}
\alias{NMRMetab_plot_raw_with_bins}
\title{plot NMR_data with bins}
\usage{
NMRMetab_plot_raw_with_bins(
  raw_data,
  pattern_file = NULL,
  min_x,
  max_x,
  ymax = 1e+06
)
}
\arguments{
\item{raw_data}{a data.frame. Column as variable and rows as sample}

\item{pattern_file}{pattern file with bins}

\item{min_x}{double}

\item{max_x}{double}

\item{ymax}{int height of the rectangle and of the text for the name of the bin}
}
\description{
plot binned spectra
}
\author{
Michele Fresneda Alarcon
}
