% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRMetab_PLSDA_average_prediction.R
\name{NMRMetab_PLSDA_average_prediction}
\alias{NMRMetab_PLSDA_average_prediction}
\alias{average_prediction_metrics}
\title{NMRMetab_average_prediction}
\usage{
average_prediction_metrics(
  dat,
  groupID = "disease_status",
  index_col = 9,
  components_model = 5,
  iterations = 5
)
}
\arguments{
\item{dat}{a data.frame. Column as variable and rows as sample}

\item{groupID}{a data.frame. with rows as metabolites.. column names are 'HMDB' and 'Metab'}

\item{components_model}{boolean. save all into result into file}

\item{iterations}{something}
}
\description{
This function will take in either a object of class data.frame and a using a second dataframe with a list of all the metabolite identifed, will return the CRS score for all the bins. this could work with other metabolite identifiers; for example bins with the name of the     metabolite, as  log as they are unique. mistakes could happen when greple take NAD and NADH
}
\author{
Michele Fresneda Alarcon
}
