% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRMetab_S2N.R
\name{NMRmetab_S2N}
\alias{NMRmetab_S2N}
\title{Signal to Noise calculation}
\usage{
NMRmetab_S2N(data, index_column, pattern_file, noise_column)
}
\arguments{
\item{data}{a data.frame. Column as samples and rows as ppm. fits columm 'ppm'. this come from bruker topspin}

\item{pattern_file}{a data.frame with column as 'min_ppm', max_ppm' and 'bin'. this is the pattern file the you created}
}
\description{
this function will calculate the signal to noise ration for each sample and each bin.
}
\author{
Michele Fresneda Alarcon
}
