% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/NMRMetab_readBruker.R
\name{NMRMmetab_readBruker}
\alias{NMRMmetab_readBruker}
\title{read Bruker files}
\usage{
NMRMmetab_readBruker(BrukerDataDir)
=======
% Please edit documentation in R/NMRmetab_readBruker.R
\name{NMRMmetab_readBruker}
\alias{NMRMmetab_readBruker}
\alias{readBruker}
\title{read Bruker files}
\usage{
readBruker(BrukerDataDir)
>>>>>>> b1f0e94cf480a7b6e5e9136fbc6f6428acf4b709
}
\arguments{
\item{BrukerDataDir}{a string. path to the NMR experiment folder}
}
\description{
takes the path where al the experiment files are, returns a dataframe with ppm as first column and samples as subsequent column
}
\author{
Dr. Jie Hao

Michele Fresneda Alarcon
}
