% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRMetab_binning.R
\name{NMRMetab_binning}
\alias{NMRMetab_binning}
\title{Binning of NMR spectra}
\usage{
NMRMetab_binning(data, pattern)
}
\arguments{
\item{data}{a data.frame. Column as samples and rows as ppm. fits columm 'ppm'. this come from bruker topspin}

\item{pattern}{a data.frame with column as 'min_ppm', max_ppm' and 'bin'. this is the pattern file the you created}
}
\description{
using a pattern file and the data extracted from bruker experiment files. this function sues the bins min and max to get a value, area under the curve to attribute to tat particular bin
}
\author{
Michele Fresneda Alarcon
}
