% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRMetab_foldchange.R
\name{NMRMetab_foldchange}
\alias{NMRMetab_foldchange}
\title{foldchange function}
\usage{
NMRMetab_foldchange(data, groupID, index_col = 3, dividendID, divisorID)
}
\arguments{
\item{data}{a data.frame. Column as variable and rows as sample}

\item{groupID}{string. name off the variable containing your grouping details must only have 2 facors eg Healthy vs Diseases}

\item{index_col}{index colum oif first mobservation}

\item{dividendID}{string. top of the division}

\item{divisorID}{string. what group you're dividing by}
}
\description{
This function will take in either a object of class data.frame and a using a second dataframe with a list of all the metabolite identifed, will return the CRS score for all the bins. this could work with other metabolite identifiers; for example bins with the name of the     metabolite, as  log as they are unique. mistakes could happen when greple take NAD and NADH
}
\author{
Michele Fresneda Alarcon
}
