% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRMetab_CRS.R
\name{NMRMetab_names_to_HMDB}
\alias{NMRMetab_names_to_HMDB}
\title{swap metabolite names to HMDB}
\usage{
NMRMetab_names_to_HMDB(dat, metabolite_list, index_col = 3)
}
\arguments{
\item{dat}{a data.frame. Column as variable and rows as sample}

\item{metabolite_list}{a data.frame. with rows as metabolites.. column names are 'HMDB' and 'Metab'}

\item{index_col}{column number with first metabolite}
}
\description{
swap metabolite names to HMDB
}
\author{
Michele Fresneda Alarcon
}
