% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRMetab_PLSDA_average_prediction.R
\name{NMRMetab_PLSDA_average_prediction}
\alias{NMRMetab_PLSDA_average_prediction}
\alias{average_prediction_metrics}
\title{NMRMetab_average_prediction}
\usage{
average_prediction_metrics(
  dat,
  groupID = "disease_status",
  index_col = 9,
  components_model = 5,
  iterations = 5
)
}
\arguments{
\item{dat}{a data.frame. Column as variable and rows as sample. data already nomalised}

\item{groupID}{a string. name of the column with the groupings}

\item{components_model}{integer. number of maximum componets in the model}

\item{iterations}{repetitions}
}
\description{
repeat the train and test split, vip extraction and prediction over number of iterations. to extract average prediction metrics
}
\author{
Michele Fresneda Alarcon
}
