% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRMetab_readBruker.R
\name{NMRMmetab_readBruker}
\alias{NMRMmetab_readBruker}
\title{read Bruker files}
\usage{
NMRMmetab_readBruker(BrukerDataDir)
}
\arguments{
\item{BrukerDataDir}{a string. path to the NMR experiment folder}
}
\description{
takes the path where al the experiment files are, returns a dataframe with ppm as first column and samples as subsequent column
}
\author{
Dr. Jie Hao

Michele Fresneda Alarcon
}
